+++ src/modules/kiosks/Details.js
@@ -16,7 +16,7 @@ import {
   getKioskSingle,
   getKioskShelves,
   getOrgIdFromKiosk,
-  getOrgName,
+  getOrgData,
 } from './selectors';
 import {
   resetKiosk,
@@ -53,7 +53,7 @@ const KioskDetails = ({
   getKiosk,
   orgId,
   getOrganizationById,
-  orgName,
+  orgData,
   rootUser,
 }) => {
   const { id } = match.params;
@@ -85,7 +85,7 @@ const KioskDetails = ({
       createRefill(kiosk._id);
     }
   };
-  const loaded = !isKioskLoading && orgName;
+  const loaded = !isKioskLoading && orgData;
   return loaded ? (
     <>
       <Grid stackable>
@@ -97,7 +97,7 @@ const KioskDetails = ({
                   <Breadcrumbs
                     backLink={backLink}
                     links={links}
-                    activeLink={kiosk.name}
+                    activeLink={kiosk && kiosk.name}
                   />
                 </Segment>
               </Grid.Column>
@@ -106,20 +106,20 @@ const KioskDetails = ({
               <Grid.Column stretched>
                 <Segment>
                   <DetailsHeader
-                    name={kiosk.name}
-                    temp={kiosk.temperature.value}
-                    doorStatus={kiosk.doorStatus}
-                    temperature={kiosk.temperature}
-                    session={kiosk.session}
+                    name={kiosk && kiosk.name}
+                    temp={kiosk && kiosk.temperature.value}
+                    doorStatus={kiosk && kiosk.doorStatus}
+                    temperature={kiosk && kiosk.temperature}
+                    session={kiosk && kiosk.session}
                   />
                   <Divider />
                   <DetailsInfo
-                    serial={`#${kiosk.serialNumber}`}
-                    session={kiosk.session}
-                    location={kiosk.location}
-                    ownerOrganization={orgName}
-                    notes={kiosk.notes}
-                    pin={kiosk.pin}
+                    serial={`#${kiosk && kiosk.serialNumber}`}
+                    session={kiosk && kiosk.session}
+                    location={kiosk && kiosk.location}
+                    ownerOrganization={orgData.name}
+                    notes={kiosk && kiosk.notes}
+                    pin={kiosk && kiosk.pin}
                   >
                     <>
                       <CustomButton
@@ -141,21 +141,23 @@ const KioskDetails = ({
                         icon="thermometer quarter"
                         label="Temp. Log"
                         onClick={() =>
-                          history.push(`/kiosks/log/temp/${kiosk._id}`)
+                          history.push(`/kiosks/log/temp/${kiosk && kiosk._id}`)
                         }
                       />
                       <CustomButton
                         icon="line graph"
                         label="Activity Log"
                         onClick={() =>
-                          history.push(`/kiosks/log/activity/${kiosk._id}`)
+                          history.push(
+                            `/kiosks/log/activity/${kiosk && kiosk._id}`,
+                          )
                         }
                       />
                       <CustomButton
                         icon="setting"
                         label="Settings"
                         onClick={() =>
-                          history.push(`/kiosks/settings/${kiosk._id}`)
+                          history.push(`/kiosks/settings/${kiosk && kiosk._id}`)
                         }
                       />
                     </>
@@ -167,7 +169,7 @@ const KioskDetails = ({
               <Grid.Column>
                 <DetailsLoadCells
                   cells={loadCells.list}
-                  kioskName={kiosk.name}
+                  kioskName={kiosk && kiosk.name}
                   currentKioskSide={currentKioskSide}
                 />
               </Grid.Column>
@@ -178,7 +180,9 @@ const KioskDetails = ({
         <Grid.Column width={5}>
           <Grid>
             <DetailQRCode
-              qrCode={`http://qrdeeplink.livello.com?id=${kiosk.qrcode}`}
+              qrCode={`http://qrdeeplink.livello.com?qrCode=${kiosk &&
+                kiosk.qrcode}&slug=${orgData &&
+                orgData.slug}&appleId=${orgData && orgData.appleId}`}
             />
             <Grid.Row>
               <Grid.Column>
@@ -198,7 +202,7 @@ const mapStateToProps = state => ({
   kiosk: getKioskSingle(state),
   loadCells: getKioskShelves(state),
   orgId: getOrgIdFromKiosk(state),
-  orgName: getOrgName(state),
+  orgData: getOrgData(state),
   isKioskLoading: state.kiosks.isKioskLoading,
   currentKioskSide: state.kiosks.currentKioskSide,
   rootUser: state.user.root,