+++ src/modules/kiosks/components/CustomizeGeneral.js
@@ -1,9 +1,17 @@
 import React, { useEffect, useState } from 'react';
 import { useDispatch, connect } from 'react-redux';
-import { Grid, Form, Button, FormRadio, Popup, Icon } from 'semantic-ui-react';
+import {
+  Grid,
+  Form,
+  Button,
+  FormRadio,
+  Popup,
+  Icon,
+  Header,
+  Divider,
+} from 'semantic-ui-react';
 import { Formik, Field } from 'formik';
 import * as Yup from 'yup';
-import prettierNumber from 'lib/prettierNumber';
 import FormInput from 'modules/shared/components/FormInput';
 import FormSelect from 'modules/shared/components/FormSelect';
 import FormCheckbox from 'modules/shared/components/FormCheckbox';
@@ -116,18 +124,9 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
   useEffect(() => {
     if (kioskProps.minimumAge === '') setAge('0');
     else setAge(kioskProps.minimumAge.toString());
-    if (!type || type === '') setType(kioskProps.paymentType);
-    setMemberCard(kioskProps.memberCardEnabled);
-  }, []);
-
-  useEffect(() => {
-    if (kioskProps.minimumAge === '') setAge('0');
-    else setAge(kioskProps.minimumAge.toString());
     setType(kioskProps.paymentType);
     setMemberCard(kioskProps.memberCardEnabled);
     setServiceCheckEnabled(serviceCheckEnabled);
-    kioskProps.technicianPin = kioskProps.technicianPin.toString();
-    // kioskProps.pin = kioskProps.pin.toString();
   }, [kioskProps]);
 
   const languages = [
@@ -166,49 +165,60 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
     },
   ];
 
-  // const TimeOptions = () => {
-  //   let time = [];
-  //   for (let i = 0; i <= 23; i++) {
-  //     time.push({ key: i, value: i, text: `${String(i).padStart(2, '0')}:00` });
-  //   }
-  //   return time;
-  // };
-  console.log(kioskProps);
   return (
     <Formik
       initialValues={kioskProps}
       onSubmit={onSubmit}
       enableReinitialize
+      validateOnChange
       validationSchema={Yup.object().shape({
         preAuth: Yup.number().max(50, 'Amount exceeds € 50.'),
-        technicianPin: Yup.number()
-          .notOneOf(
-            [Yup.ref('pin')],
-            "Replenishment PIN and Technician PIN can't be the same.",
-          )
-          .test(
-            'len',
-            'Must be exactly 4 digits',
-            val => val.toString().length === 4,
-          )
-          .positive('Must be greater than zero')
-          .typeError('Invalid Input: Numbers please'),
         pin: Yup.number()
-          .notOneOf(
-            [Yup.ref('technicianPin')],
-            "Replenishment PIN and Technician PIN can't be the same.",
-          )
+          .required('This field is required')
+          .test({
+            name: 'check-same-pin',
+            test: function(val) {
+              if (val && val === this.parent.technicianPin) {
+                return this.createError({
+                  path: 'pin',
+                  message:
+                    "Replenishment PIN and Technician PIN can't be the same",
+                });
+              } else return true;
+            },
+          })
+          .typeError('Invalid Input: Numbers please')
+          .positive('Must be greater than zero')
           .test(
             'len',
             'Must be exactly 4 digits',
-            val => val.toString().length === 4,
-          )
+            val => val && val.toString().length === 4,
+          ),
+        technicianPin: Yup.number()
+          .required('This field is required')
+          .test({
+            name: 'check-same-pin',
+            test: function(val) {
+              if (val && val === this.parent.pin) {
+                return this.createError({
+                  path: 'technicianPin',
+                  message:
+                    "Replenishment PIN and Technician PIN can't be the same",
+                });
+              } else return true;
+            },
+          })
+          .typeError('Invalid Input: Numbers please')
           .positive('Must be greater than zero')
-          .typeError('Invalid Input: Numbers please'),
+          .test(
+            'len',
+            'Must be exactly 4 digits',
+            val => val && val.toString().length === 4,
+          ),
       })}
     >
       {({ dirty, handleSubmit, resetForm, setFieldValue }) => (
-        <Form onSubmit={handleSubmit}>
+        <Form onSubmit={handleSubmit} style={{ marginTop: '30px' }}>
           <Grid>
             <Grid.Row columns="equal">
               <Grid.Column>
@@ -381,7 +391,6 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
                   name="pin"
                   label="Replenishment PIN"
                   required
-                  placeholder="1234"
                   widthLimit
                   component={FormInput}
                 />
@@ -391,7 +400,6 @@ const CustomizeScreen = ({ cancelHandler, kioskProps, kiosk }) => {
                   name="technicianPin"
                   label="Technician PIN"
                   required
-                  placeholder="5678"
                   widthLimit
                   component={FormInput}
                 />