+++ src/modules/kiosks/components/DetailsInfo.js
@@ -8,7 +8,7 @@ const DetailsInfo = ({
   children,
   session,
   serial,
-  location: { address },
+  location,
   ownerOrganization,
   notes,
   pin,
@@ -18,10 +18,10 @@ const DetailsInfo = ({
   let addressLine3 = '';
   let addressSecondLine = '';
   let addressThirdLine = '';
-  // address
-  //   ? [address.city, address.line1, address.state].filter(el => el).join(', ')
-  //   : '';
-  if (address.line1 !== null) {
+
+  let address = location && location.address;
+
+  if (address && address.line1 !== null) {
     addressLine1 += address.name ? `${address.name} \n ` : '';
     addressLine2 = `${address.line1}`;
     addressLine3 += address.line2 ? `${address.line2}` : '';